requirements.txt
chromadb>=1.0.0
fastapi>=0.115.0
google-genai>=1.0.0
Pillow>=11.0.0
pdfplumber>=0.11.0
pymupdf4llm
python-dotenv>=1.0.1
uvicorn[standard]>=0.34.0
